Style animation key button
