+++ src/components/animation-app.scss
@@ -27,6 +27,7 @@
 .key-box.animation {
   display: flex;
   flex-direction: column;
+  align-items: center;
 
   background-color: #f9efff;
   border-radius: 14px;
