+++ src/components/video-controls.scss
@@ -18,10 +18,6 @@
       border-radius: 14px;
     }
 
-    .button-icon {
-      fill: #aa2ff0;
-    }
-
     .rc-slider-handle {
       border: 2px solid #aa2ff0;
     }
@@ -42,78 +38,12 @@
       }
     }
 
-    .button-icon {
-      fill: #0481a0;
-    }
-
     .rc-slider-handle {
       border: 2px solid #0481a0;
     }
 
     .rc-slider-dot {
       background-color: #0481a0;
-    }
-  }
-
-  .video-view-button {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    cursor: pointer;
-
-    height: 32px;
-    margin: 0 8px;
-    padding-left: 0px;
-
-    .button-icon-container {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 20px;
-      
-      .button-icon {
-        position: relative;
-
-        &.play-button-icon {
-          left: 4px;
-          top: -1px;
-        }
-        &.pause-button-icon {
-          left: 2px;
-          top: -1px;
-        }
-      }
-    }
-
-    .button-label {
-      display: flex;
-      justify-content: center;
-      width: 40px;
-    }
-
-    &.animation {
-      background-color: #ecc9ff;
-      border: 2px solid #aa2ff0;
-      border-radius: 16px;
-      
-      &:hover {
-        background-color: #e3b1ff;
-      }
-      &:active {
-        background-color: #d68cff;
-      }
-    }
-    &.simulation {
-      background-color: #b7e2ec;
-      border: 2px solid #0481a0;
-      border-radius: 5px;
-      
-      &:hover {
-        background-color: #93d5e4;
-      }
-      &:active {
-        background-color: #6fc6da;
-      }
     }
   }
